requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.3.1
django-filter==23.5
django-hvad==1.8.0
django-jazzmin==2.6.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
